refactor: start menu
-Refactor start menu into smaller components. -Add styling to elements.
